Refactor test helpers - move test helpers to main - add support for inline golden values - add test for `flux --version`
