+++ cmd/flux/testdata/trace/deployment.txt
@@ -23,4 +23,3 @@ URL:            https://stefanprodan.github.io/podinfo
 Revision:       8411f23d07d3701f0e96e7d9e503b7936d7e1d56
 Status:         Last reconciled at 2021-07-11 00:25:46 +0000 UTC
 Message:        Fetched revision: 8411f23d07d3701f0e96e7d9e503b7936d7e1d56
-
