--- cmd/flux/testdata/trace/no-args.txt
@@ -1,1 +1,0 @@
-object name is required
